requirements_mtk.txt
matplotlib
opencv-python
paho-mqtt
tqdm
pybind11
mypy
torch
moviepy<2.0.0
pandas
scipy